+++ csv/csv.go
@@ -10,8 +10,10 @@ package csv
 
 import (
 	"bufio"
+	"bytes"
 	"errors"
 	"io"
+	"strconv"
 	"strings"
 )
 
@@ -19,9 +21,16 @@ import (
 
 // Reader is CSV reader struct
 type Reader struct {
-	Comma rune
-	br    *bufio.Reader
+	Comma      rune
+	SkipHeader bool
+
+	headerSkipped bool
+
+	br *bufio.Reader
 }
+
+// Row is CSV row
+type Row []string
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -36,17 +45,24 @@ var ErrNilReader = errors.New("Reader is nil")
 // NewReader create new CSV reader
 func NewReader(r io.Reader) *Reader {
 	return &Reader{
-		Comma: ';',
-		br:    bufio.NewReader(r),
+		Comma:      ';',
+		SkipHeader: false,
+
+		br: bufio.NewReader(r),
 	}
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // Read reads line from CSV file
-func (r *Reader) Read() ([]string, error) {
+func (r *Reader) Read() (Row, error) {
 	if r == nil {
 		return nil, ErrNilReader
+	}
+
+	if r.SkipHeader && !r.headerSkipped {
+		r.br.ReadLine()
+		r.headerSkipped = true
 	}
 
 	str, _, err := r.br.ReadLine()
@@ -59,13 +75,18 @@ func (r *Reader) Read() ([]string, error) {
 }
 
 // ReadTo reads data to given slice
-func (r *Reader) ReadTo(dst []string) error {
+func (r *Reader) ReadTo(dst Row) error {
 	if r == nil {
 		return ErrNilReader
 	}
 
 	if len(dst) == 0 {
 		return ErrEmptyDest
+	}
+
+	if r.SkipHeader && !r.headerSkipped {
+		r.br.ReadLine()
+		r.headerSkipped = true
 	}
 
 	str, _, err := r.br.ReadLine()
@@ -90,10 +111,128 @@ func (r *Reader) WithComma(comma rune) *Reader {
 	return r
 }
 
+// WithHeaderSkip sets header skip flag
+func (r *Reader) WithHeaderSkip(flag bool) *Reader {
+	if r == nil {
+		return nil
+	}
+
+	r.SkipHeader = flag
+
+	return r
+}
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// parseAndFill parses line
-func parseAndFill(src string, dst []string, sep string) {
+// Size returns size of the row
+func (r Row) Size() int {
+	return len(r)
+}
+
+// Cells returns number of cells filled with data
+func (r Row) Cells() int {
+	var size int
+
+	for _, c := range r {
+		if len(c) > 0 {
+			size++
+		}
+	}
+
+	return size
+}
+
+// IsEmpty returns true if all cells are empty
+func (r Row) IsEmpty() bool {
+	return r.Cells() == 0
+}
+
+// Has returns true if row contains cell with given index filled with data
+func (r Row) Has(index int) bool {
+	return index < len(r) && r[index] != ""
+}
+
+// Get returns value of the cell with given index
+func (r Row) Get(index int) string {
+	if index >= len(r) {
+		return ""
+	}
+
+	return r[index]
+}
+
+// GetB returns cell value as boolean
+func (r Row) GetB(index int) bool {
+	switch strings.ToLower(r.Get(index)) {
+	case "1", "true", "t", "yes", "y", "+":
+		return true
+	}
+
+	return false
+}
+
+// GetI returns cell value as int
+func (r Row) GetI(index int) (int, error) {
+	return strconv.Atoi(r.Get(index))
+}
+
+// GetF returns cell value as float
+func (r Row) GetF(index int) (float64, error) {
+	return strconv.ParseFloat(r.Get(index), 10)
+}
+
+// GetU returns cell value as uint64
+func (r Row) GetU(index int) (uint64, error) {
+	return strconv.ParseUint(r.Get(index), 10, 64)
+}
+
+// ForEach executes given function for every cell in a row
+func (r Row) ForEach(fn func(index int, value string) error) error {
+	for i, c := range r {
+		err := fn(i, c)
+
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// ToString returns string representation of row
+func (r Row) ToString(comma rune) string {
+	var buf bytes.Buffer
+
+	for i, c := range r {
+		buf.WriteString(c)
+
+		if i+1 != len(r) {
+			buf.WriteRune(comma)
+		}
+	}
+
+	return buf.String()
+}
+
+// ToBytes returns representation of row as a byte slice
+func (r Row) ToBytes(comma rune) []byte {
+	var buf bytes.Buffer
+
+	for i, c := range r {
+		buf.WriteString(c)
+
+		if i+1 != len(r) {
+			buf.WriteRune(comma)
+		}
+	}
+
+	return buf.Bytes()
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// parseAndFill parses CSV row and fills slice with data
+func parseAndFill(src string, dst Row, sep string) {
 	l := len(dst)
 
 	if src == "" {
@@ -124,12 +263,12 @@ func parseAndFill(src string, dst []string, sep string) {
 	}
 
 	if i < l-1 {
-		clean(dst, i, l)
+		clean(dst, i+1, l)
 	}
 }
 
 // clean cleans destination slice
-func clean(dst []string, since, to int) {
+func clean(dst Row, since, to int) {
 	for i := since; i < to; i++ {
 		dst[i] = ""
 	}