+++ csv/csv_test.go
@@ -8,6 +8,7 @@ package csv
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
+	"errors"
 	"io"
 	"os"
 	"testing"
@@ -17,9 +18,11 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-const _DATA = `123,ABC,A_C,A C,
-123,ABC,
-123,ABC,A_C,A C,123,ABC,A_C,A C
+const _DATA = `ID,FIRST NAME,LAST NAME,BALANCE
+1,John,Doe,0.34
+2,Fiammetta,Miriana,30
+3,Mathew,Timothy,34.19371,1,2
+4,Lou
 `
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -56,29 +59,29 @@ func (s *CSVSuite) TestRead(c *C) {
 
 	defer fd.Close()
 
-	reader := NewReader(fd).WithComma(',')
-	count := 0
+	line := 0
+	reader := NewReader(fd).WithComma(',').WithHeaderSkip(true)
 
 	for {
-		rec, err := reader.Read()
+		row, err := reader.Read()
 
 		if err == io.EOF {
 			break
 		}
 
-		switch count {
+		switch line {
 		case 0:
-			c.Assert(rec, HasLen, 5)
-			c.Assert(rec, DeepEquals, []string{"123", "ABC", "A_C", "A C", ""})
+			c.Assert(row, HasLen, 4)
+			c.Assert(row, DeepEquals, Row{"1", "John", "Doe", "0.34"})
 		case 1:
-			c.Assert(rec, HasLen, 3)
-			c.Assert(rec, DeepEquals, []string{"123", "ABC", ""})
+			c.Assert(row, HasLen, 4)
+			c.Assert(row, DeepEquals, Row{"2", "Fiammetta", "Miriana", "30"})
 		case 2:
-			c.Assert(rec, HasLen, 8)
-			c.Assert(rec, DeepEquals, []string{"123", "ABC", "A_C", "A C", "123", "ABC", "A_C", "A C"})
+			c.Assert(row, HasLen, 6)
+			c.Assert(row, DeepEquals, Row{"3", "Mathew", "Timothy", "34.19371", "1", "2"})
 		}
 
-		count++
+		line++
 	}
 }
 
@@ -90,19 +93,17 @@ func (s *CSVSuite) TestReadTo(c *C) {
 
 	defer fd.Close()
 
-	reader := NewReader(fd)
-	reader.Comma = ','
+	line := 0
+	reader := NewReader(fd).WithComma(',').WithHeaderSkip(true)
 
-	count := 0
-
-	var rec []string
+	var row Row
 
 	for {
-		err := reader.ReadTo(rec)
+		err := reader.ReadTo(row)
 
-		if rec == nil {
+		if row == nil {
 			c.Assert(err, NotNil)
-			rec = make([]string, 8)
+			row = make(Row, 4)
 			continue
 		}
 
@@ -110,41 +111,79 @@ func (s *CSVSuite) TestReadTo(c *C) {
 			break
 		}
 
-		switch count {
+		switch line {
 		case 0:
-			c.Assert(rec, DeepEquals, []string{"123", "ABC", "A_C", "A C", "", "", "", ""})
+			c.Assert(row, DeepEquals, Row{"1", "John", "Doe", "0.34"})
 		case 1:
-			c.Assert(rec, DeepEquals, []string{"123", "ABC", "", "", "", "", "", ""})
+			c.Assert(row, DeepEquals, Row{"2", "Fiammetta", "Miriana", "30"})
 		case 2:
-			c.Assert(rec, DeepEquals, []string{"123", "ABC", "A_C", "A C", "123", "ABC", "A_C", "A C"})
+			c.Assert(row, DeepEquals, Row{"3", "Mathew", "Timothy", "34.19371"})
+		case 3:
+			c.Assert(row, DeepEquals, Row{"4", "Lou", "", ""})
 		}
 
-		count++
+		line++
 	}
 }
 
 func (s *CSVSuite) TestLineParser(c *C) {
-	data := make([]string, 2)
+	data := make(Row, 2)
 
 	parseAndFill("ABCD", data, ";")
-	c.Assert(data, DeepEquals, []string{"ABCD", ""})
+	c.Assert(data, DeepEquals, Row{"ABCD", ""})
 
 	parseAndFill("", data, ";")
-	c.Assert(data, DeepEquals, []string{"", ""})
+	c.Assert(data, DeepEquals, Row{"", ""})
 
 	parseAndFill("A;B;C;D;E", data, ";")
-	c.Assert(data, DeepEquals, []string{"A", "B"})
+	c.Assert(data, DeepEquals, Row{"A", "B"})
 }
 
 func (s *CSVSuite) TestNil(c *C) {
 	var r *Reader
-	var b = []string{""}
+	var b = Row{""}
 
 	_, err := r.Read()
 
 	c.Assert(err, DeepEquals, ErrNilReader)
 	c.Assert(r.ReadTo(b), DeepEquals, ErrNilReader)
 	c.Assert(r.WithComma('X'), IsNil)
+	c.Assert(r.WithHeaderSkip(false), IsNil)
+}
+
+func (s *CSVSuite) TestRow(c *C) {
+	r := Row{"test", "", "1234", "12.34", "Yes"}
+	c.Assert(r.Size(), Equals, 5)
+	c.Assert(r.Cells(), Equals, 4)
+	c.Assert(r.IsEmpty(), Equals, false)
+	c.Assert(r.Has(0), Equals, true)
+	c.Assert(r.Has(1), Equals, false)
+	c.Assert(r.Get(0), Equals, "test")
+	c.Assert(r.Get(10), Equals, "")
+
+	ci, err := r.GetI(2)
+	c.Assert(ci, Equals, 1234)
+	c.Assert(err, IsNil)
+
+	cf, err := r.GetF(3)
+	c.Assert(cf, Equals, 12.34)
+	c.Assert(err, IsNil)
+
+	cu, err := r.GetU(2)
+	c.Assert(cu, Equals, uint64(1234))
+	c.Assert(err, IsNil)
+
+	c.Assert(r.GetB(1), Equals, false)
+	c.Assert(r.GetB(4), Equals, true)
+
+	dummyFn := func(i int, v string) error { return nil }
+	c.Assert(r.ForEach(dummyFn), IsNil)
+
+	dummyFn = func(i int, v string) error { return errors.New("1") }
+	c.Assert(r.ForEach(dummyFn), NotNil)
+
+	c.Assert(r.ToString(';'), Equals, "test;;1234;12.34;Yes")
+	c.Assert(string(r.ToBytes(';')), Equals, "test;;1234;12.34;Yes")
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //