+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 12.121.0
 
 - `[initsystem/sdnotify]` Added new package for sending messages to systemd
+- `[csv]` Added helpers for working with CSV row
+- `[csv]` Added option to skip header
 - `[support/deps]` Updated for compatibility with the latest version of [depsy](https://kaos.sh/depsy)
 - `[terminal/tty]` Improved check for systemd